refactor to better project layout
