+++ cfg/1.1.0/domain-controller.yaml
@@ -1,9 +1,9 @@
 ---
 controls:
-version: 1.10
+version: 1.1.0
 id: 1
 description: "Initial Setup"  
-type: "master"
+type: "domain-controller"
 groups:
 - id: 1.1
   description: "Password Policy"
