+++ check/check.go
@@ -1,0 +1,1 @@
+package check
